Update readme
Adding slightly increased clarity about VOCs and grammar adjustments. The term VOCs is thrown around a lot (especially in Discord) with the intent of harmful chemicals but VOCs are also the harmless chemicals such as typical odors like farts. I make the mistake myself a lot by using VOCs to refer to the harmful components while forgetting of the harmless ones. This was especially apparent in our discussions on VOC sensors.